README and Manifests
This commit adds a README which includes an overview, a gettting started guide,
and a contribution section. The audit-webhook manifest has been updated to use a
local PersistentVolume and restrict which nodes it is scheduled to. The
audit-config manifest has been updated to use the DNS name instead of a
ClusterIP.

Signed-off-by: Dhruv Jain <2dhruv@gmail.com>
Co-authored-by: Stan Wong <swong394@gmail.com>